// src/generated/prisma/enums.ts
export const CitationType = {
  OFFICIAL: 'OFFICIAL',
  MEDIA: 'MEDIA',
  QA: 'QA',
  FORUM: 'FORUM',
  WIKI: 'WIKI',
  MAP: 'MAP',
  LOCAL_LIFE: 'LOCAL_LIFE',
  SOCIAL: 'SOCIAL',
  UNKNOWN: 'UNKNOWN',
  OTHER: 'OTHER'
} as const

export type CitationType = (typeof CitationType)[keyof typeof CitationType]